+++ retex/renderer-desktop/src/main/java/com/himamis/retex/renderer/desktop/graphics/Graphics2DD.java
@@ -114,8 +114,8 @@ public class Graphics2DD implements Graphics2DInterface {
 	}
 
 	@Override
-	public void fillRect(int x, int y, int width, int height) {
-		impl.fillRect(x, y, width, height);
+	public void fillRect(double x, double y, double width, double height) {
+		impl.fillRect((int) x, (int) y, (int) width, (int) height);
 	}
 
 	@Override
